feat: full code revamp
Due to the difficulty of correcting the bugs issued from the
refactorization; due to the "crossed relations", Plane(Passenger(Plane))
many method and attribute calls where not properly automatically
refactored. In consequence, a full revision of the code has been
executed in order to ensure proper execution with a "renewed" and more
clear structure.

related: #2