fix: correction of the action space definition
The action space definition was NOT properly defined. The multidiscrete
object allows you to define a array, each element indicating a
different action, and a value for each element (indicating the number
of options in each action). In this development, as we try to create a
__queue system__ each position of the queue can have any value of the
queue order; resulting in an array of:
   [num_passengers * num_passengers]

related: #3