explore_boarding (#6)
* chore: update environment.yml

* feat: onboard simulator adaptation

The adaptation is a process of taking the found code:
- Document methods and functions
- Refactor names to make them more self-explanatory
- Remove boilerplate, visualization, not required in this step

related: #1

* chore: add initialization of the folder

* feat: full code revamp

Due to the difficulty of correcting the bugs issued from the
refactorization; due to the "crossed relations", Plane(Passenger(Plane))
many method and attribute calls where not properly automatically
refactored. In consequence, a full revision of the code has been
executed in order to ensure proper execution with a "renewed" and more
clear structure.

related: #2

* refactor: unified the theme ofr each file

Before some methods where present in the simulation file, now due to the
usage in the gym environment too; they have been moved to the 'utils'
file. Now the methods and clases segmentation is much more clear

related: #2

* feat: creation of the OpenAI gym environment

Creation of the basic OpenAI gym's environment, following the
implementation support found on:
- https://towardsdatascience.com/creating-a-custom-openai-gym-environment-for-stock-trading-be532be3910e
- https://ai-mrkogao.github.io/reinforcement%20learning/openaigymtutorial/
- https://towardsdatascience.com/reinforcement-learning-with-openai-d445c2c687d2

Testing is not yet implemented

related: #3

* fix: correction of the action space definition

The action space definition was NOT properly defined. The multidiscrete
object allows you to define a array, each element indicating a
different action, and a value for each element (indicating the number
of options in each action). In this development, as we try to create a
__queue system__ each position of the queue can have any value of the
queue order; resulting in an array of:
   [num_passengers * num_passengers]

related: #3

* fix: correction of the observable space definition

The action observable definition was NOT properly defined. The
multidiscrete object allows you to define a array, each element
indicating a different action, and a value for each element (indicating
the number of options in each action). In this development, as we try
to create a __queue system__ each position of the queue can have any
value of the queue order.

related: #3

* fix: correction of the observation space initialization

Fix required to properly append the passenger's baggage value, to the
rest of information defining it.

related: #3

* fix: small fix related to definition of a return value by the step
method

related: #3

* fix: small fix related to the rewards evaluation

As the algorithm needs to learn to maximize the reward, the less time
the simulation spends, the better the queue has been prioritized

related: #3

* feat: creation testing code for debugging purposes

related: #3

* chore: reordering of files to match OpenAI gym' structure

related: #3

* feat: defined OpenAI gym's setup

related: #3

* feat: defined OpenAI gym's structure

related: #3

* feat: gym environment proper definition

Renaming and restructuring of the different components from the Gym
environment, in order to follow OpenAI's instructions.

realted: #3

* feat: gym environment proper definition

Renaming and restructuring of the different components from the Gym
environment, in order to follow OpenAI's instructions.

realted: #3

* feat: gym environment tester updated

Refactoring and updated of the gym environment script

close: #3